pre-commit 自动更新发布日期 pubDatetime Fix #60
