+++ README.md
@@ -1,26 +1,3 @@
 # WTTO'S BLOG
 
-## TODO
-
-- [x] ~~文章显示 code 行内样式修改~~
-- [x] ~~侧边的菜单导航还要再测试更多级别的展示~~
-- [x] ~~侧边导航滚动页面的时候自动选中~~
-- [x] ~~阅读时的回到顶部按钮~~
-- [x] ~~文章在宽屏时，展示一张二维码图片，可以让用户手机扫码在移动端观看~~
-- [x] ~~Github CI 部署~~
-- [x] ~~PUBLIC_GOOGLE_SITE_VERIFICATION 配置~~
-- [x] ~~去掉作者字段~~
-- [ ] pre-commit 自动更新/创建相关联的 issue
-- [ ] pre-commit 自动更新发布日期 pubDatetime，注意 draft 字段
-- [ ] 文章尾部添加版权说明，参考 [luolei's blog](https://luolei.org/what-i-use-chrome-extension-2023#chrome-%E6%8F%92%E4%BB%B6-%E5%BC%80%E5%8F%91%E8%B0%83%E8%AF%95)
-- [ ] 文章链接到 Github 源文件以及相关联的 issue
-- [ ] 归档按照日期，只显示标题，单页显示全部
-- [ ] tab-index 各个页面整理一遍
-- [ ] 文章目录，总结整理一遍
-- [ ] 升级框架版本
-- [ ] 跳转页面太慢了，添加 ProgressBar
-- [ ] 评论插件 (可放到下期)
-
-## Bug
-
-- [ ] prose 样式没有处理自定义的--uno 样式
+使用 [Astro](https://astro.build/) + [solidjs](https://www.solidjs.com/) 构建
